pyproject.toml
[tool.poetry]
name = "devchat"
version = "0.2.10"
description = "DevChat is an open-source tool that helps developers write prompts to generate code and documentation."
authors = ["DevChat Team <hello@devchat.ai>"]
license = "Apache License 2.0"
readme = "README.md"
homepage = "https://github.com/devchat-ai/devchat"
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Information Technology",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development"
]

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "1.10.14"
networkx = "^3.1"
openai = "^1.0rc"
rich_click = "^1.6.1"
tiktoken = "^0.4.0"
tinydb = "^4.7.1"
urllib3 = "<2.0"
importlib-metadata = "^6.8.0"
gitpython = "^3.1.32"
oyaml = "^1.0"
colorama = "^0.4.6"
tenacity = "^8.2.3"
pathspec = "^0.12.1"
importlib-resources = "^6.1.1"

[tool.poetry.scripts]
devchat = "devchat._cli.main:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
ruff = "^0.4.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py38"
line-length = 100

[tool.ruff.lint]
select = [
    "E",  # Error
    "W",  # Warning
    "F",  # pyflakes
    "I",  # isort
]
fixable = ["ALL"]

